server/yolo_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==10.4.0
ultralytics==8.3.40
numpy==1.26.4
opencv-python-headless==4.10.0.84